moved snapshot resources related classes to subpackages
