+++ envoy-control-core/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/snapshot/resource/routes/RoutesMatchers.kt
@@ -1,4 +1,4 @@
-package pl.allegro.tech.servicemesh.envoycontrol.snapshot.resource
+package pl.allegro.tech.servicemesh.envoycontrol.snapshot.resource.routes
 
 import io.envoyproxy.envoy.api.v2.route.HeaderMatcher
 import pl.allegro.tech.servicemesh.envoycontrol.protocol.HttpMethod
